Provide hotfix for Neos core performance issue
This change provides a hotfix to mitigate a performance issue in the
original implementation of the Frontend Node Route Part Handler in
Neos core. When a given document node contains thousands of child nodes,
the original route part handler may end up loading all of these nodes
into memory in order to check their URI path segment.

The fix is provided as part of this package for convenience until the
issue is resolved in Neos core and has been released.